+++ src/gmp.c
@@ -105,6 +105,7 @@
 #include "manage_groups.h"
 #include "manage_notes.h"
 #include "manage_oci_image_targets.h"
+#include "manage_overrides.h"
 #include "manage_permissions.h"
 #include "manage_port_lists.h"
 #include "manage_report_configs.h"
