+++ src/manage_sql_overrides.h
@@ -1,0 +1,11 @@
+/* Copyright (C) 2026 Greenbone AG
+ *
+ * SPDX-License-Identifier: AGPL-3.0-or-later
+ */
+
+#ifndef _GVMD_MANAGE_SQL_OVERRIDES_H
+#define _GVMD_MANAGE_SQL_OVERRIDES_H
+
+#include "manage_overrides.h"
+
+#endif // not _GVMD_MANAGE_SQL_OVERRIDES_H
